refactor UndoFieldInfo
