+++ src/components/admin/ChangesViewer/types.ts
@@ -2,3 +2,10 @@ export interface ChangeItem {
     id: string;
     name: string;
 }
+
+export interface UndoFieldInfo {
+    componentId: string;
+    fieldName: string;
+    locale?: string;
+    oldValue: any;
+}
